ai-cyber-assurance-agent/requirements.txt
anthropic>=0.40.0
streamlit>=1.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
openpyxl>=3.1.0
pypdf>=3.0.0
Pillow>=10.0.0
pytesseract>=0.3.11
